// third_party/tencentdb-agent-memory-v2/MemoryCore/src/core/types.ts
export interface Logger {
  debug?: (message: string) => void;
  info: (message: string) => void;
  warn: (message: string) => void;
  error: (message: string) => void;
}